refactor: move application call to index file
